Update on saving type of coin, giving 5% bonus. And websocket ssl config.

+++ src/main/java/com/lilcoin/level/LevelInfoDTO.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 public class LevelInfoDTO {
   private Integer level;
   private String levelTitle;
-  private Long levelPrice;
+  private Double levelPrice;
 }
